requirements.txt
Flask==2.3.1
Pillow==9.5.0
Requests==2.29.0
gunicorn==20.1.0
torch==2.0.0
torchvision==0.15.1